+++ draft-js-mention-plugin/src/modifiers/addMention.js
@@ -17,11 +17,9 @@ const addMention = (
   const entityKey = contentStateWithEntity.getLastCreatedEntityKey();
 
   const currentSelectionState = editorState.getSelection();
-  const { begin, end } = getSearchText(
-    editorState,
-    currentSelectionState,
-    mentionTrigger
-  );
+  const { begin, end } = getSearchText(editorState, currentSelectionState, [
+    mentionTrigger,
+  ]);
 
   // get selection of the @mention search text
   const mentionTextSelection = currentSelectionState.merge({